Reorganized and rewrote several classes and methods
