+++ src/main/resources/assets/homegrown/lang/en_us.json
@@ -48,6 +48,6 @@
     "block.homegrown.cantaloupe_block": "Cantaloupe Block",
 
 
-    "itemGroup.homegrown.blocks": "Homegrown:Blocks",
-    "itemGroup.homegrown.items": "Homegrown:Items"
+    "itemGroup.homegrown.blocks": "Homegrown: Blocks",
+    "itemGroup.homegrown.items": "Homegrown: Items"
 }
